requirements.txt
Flask
python-dotenv
requests
Pillow
numpy
